Move filter test into main.py tests.
